Forgot the attribute for price

+++ src/main/java/com/fakebookstore/fakebookstore/models/Book.java
@@ -66,11 +66,11 @@ public class Book {
         this.description = description;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
